+++ R/tbl_survfit.R
@@ -312,7 +312,7 @@ meta_to_df_stats <- function(meta_data, inputs, estimate_type, estimate_fun,
                                  var_label = ..4)
         )
       ),
-      # table_body -----------------------------------------------------------------
+      # table_body -------------------------------------------------------------
       table_body = map2(
         .data$df_stats, .data$var_label,
         function(df_stats, var_label) {