+++ tests/testthat/test-set_gtsummary_theme.R
@@ -78,6 +78,7 @@ test_that("setting themes", {
                list("{n} / {N} ({p}%)", c("{median} ({p25} - {p75})", "{mean} ({sd})")))
 
 
+  reset_gtsummary_theme()
   theme_gtsummary_journal("jama")
 
   expect_equal(
@@ -86,6 +87,23 @@ test_that("setting themes", {
       as_tibble(col_labels = FALSE) %>%
       purrr::pluck("estimate"),
     c(NA, NA, "1.4 (-3.6 to 6.4)", "2.0 (-3.1 to 7.0)")
+  )
+
+  tbl1 <-
+    trial %>%
+    select(age, response, trt) %>%
+    tbl_summary(by = trt, missing = "no") %>%
+    add_difference() %>%
+    as_tibble(col_labels = FALSE)
+  expect_equal(
+    tbl1 %>%
+      purrr::pluck("estimate"),
+    c("-0.44 (-4.6 to 3.7)", "-4.2% (-18% to 9.9%)")
+  )
+  expect_equal(
+    tbl1 %>%
+      purrr::pluck("estimate"),
+    c("-0.44 (-4.6 to 3.7)", "-4.2% (-18% to 9.9%)")
   )
 
   reset_gtsummary_theme()
@@ -105,4 +123,5 @@ test_that("setting themes", {
   reset_gtsummary_theme()
 })
 
+reset_gtsummary_theme()