+++ NEWS.md
@@ -4,17 +4,19 @@
 
 * Added new function `add_glance_table()` as a companion to `add_glance_source_note()`. Function adds model statistics, such as R-squared, to the bottom of the model summary table.
 
+* Added new function `add_significance_stars()` adding star indicators to significant estimates, and an explanatory footnote.
+
 * Added new function `tbl_strata()`. The function aids prepares gtsummary tables stratified by one or more variables (#679)
 
 * Adding coefficient `plot()` methods for `tbl_regression()` and `tbl_uvregression()`. Function creates a forest plot of model coefficients via `GGally::ggcoef_plot()`.
-
-* Added new function `add_significance_stars()` adding star indicators to significant estimates, and an explanatory footnote.
 
 * New function `modify_fmt_fun()` has been introduced to help update the functions that format numeric columns and rows in `.x$table_body`.
 
 * New function introduced,  `modify_table_styling()`, to update printing instructions of tables. This function replaces `modify_table_header()`, which is now soft deprecated. 
 
 * Added function `add_vif()` to include variance inflation factors in `tbl_regression()` output. (#717)
+
+* Added a generic function inline_text.gtsummary() that can report results from any gtsummary table. (#398)
 
 ### New Functionality
 
@@ -37,6 +39,8 @@
 * Added default support for `brmsfit` model in `tbl_regression()` with new method function. (#751)
 
 * Korean and Icelandic language translations added for `theme_gtsummary_language()`.
+
+* Ability to merge two or more columns with `modify_table_styling(cols_merge_pattern=)` argument.
 
 * Added theme element `"pkgwide-fun:pre_conversion"`. The function specified here will be executed on the gtsummary object before it is printed or converted with the `as_gt()`, `as_flex_table()`, etc functions. (#773)
 
@@ -67,6 +71,8 @@
 * Allowing for tidyverse shortcut notation in `tbl_survfit(estimate_fun=)` specification, e.g. `tbl_survfit(estimate_fun= ~style_sigfig(.x * 100))` (#761)
 
 * The JAMA journal theme has been updated to merge the coefficient and confidence interval columns.
+
+* Updated other inline_text() functions to wrap inline_text.gtsummary()
 
 ### Bug Fixes